integrations/permissions.py
from rest_framework.permissions import BasePermission


class AdminOnlyPermission(BasePermission):
    message = "Accès réservé aux administrateurs."

    def has_permission(self, request, view):
        user = request.user
        return bool(user and user.is_authenticated and user.role == "ADMIN")
