requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
markdown>=3.5,<4.0
reportlab>=4.0,<5.0
qrcode>=7.4,<8.0
pillow>=10.0,<11.0
django-environ>=0.11,<1.0
ruff>=0.5,<1.0
black>=24.0,<25.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
